Several improvements in accross the board
* [CORE] Add API to wait child process to hit the main module after starting the execution
* [CORE] Database: Use big integer in INT fields of sqlite3 database to avoid overflows
* [CORE] Database: recreate table when counting
* [CORE] Rename files using prolog/roi/epilog conventions
* [TOOLS] chop trace: detection of previous signal handler warning and initial TTY fd detection support
* [TOOLS] Wait for main module to be executed before start event collection in chop sample, chop trace and chop-perf-invok 
* [TOOLS] Improvements on chop-marks related scripts: recursive support
* [TOOLS] Improvements on chop-marks related scripts: compiler optimization (tail end support)
* [TOOLS] Improvements on chop-marks related scripts: check begin/end points are found
* [TOOLS] Improvements on chop-marks related scripts: use the right flags in objdump for s390x
* [TOOLS] chop disasm: disassembly all executable sections, not just .text
* [TOOLS] Add initial ELF loader prototype tool
* [TOOLS] chop-perf-invok: several improvements, including support of recursive calls, and better management of the subprocess and initial preliminary support for multi-processing
* [TOOLS] Handle single invocations in clustering support tools
* [FIX] Database: Fix grouping and counting of samples to instructions, blocks and modules
* [FIX] CI: Update tests to use new database format
* [FIX] Respect external CPU affinity masks during for chop sample and chop-perf-invok
* [FIX] Fix endiandess and mask issue in s390x backend
* [FIX] chop-trace-mem: handle Microprobe generated ELF
* [MISC] Update documentation for new tools flags
* [MISC[ Update zlib version to use the latest one
* [MISC] Use standard exit codes 
* [MISC] Add logging statements across the board
* [MISC] Database: Provide more verbose information while counting instruction sample
* [MISC] Remove white space in SQL queries

+++ src/database/sql/add_event.sql
@@ -1,1 +1,1 @@
-ALTER TABLE sample ADD [{}] INTEGER;
+ALTER TABLE sample ADD [{}] BIGINT;
